services/requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu
torch==2.4.1+cpu
torchvision==0.19.1+cpu

fastapi==0.115.6
uvicorn==0.34.0
python-multipart==0.0.20
easyocr==1.7.2
Pillow==11.1.0
pydantic==2.10.4
pymongo==4.11.3
python-dotenv==1.1.0

psycopg2-binary==2.9.10
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.1